pyproject.toml
[project]
name = "regulatory-defi-compass"
version = "1.0.0"
description = "An agent that helps users assess DeFi opportunities through a compliance and regulatory lens, overlaying risk scores on top of yield data."
authors = [
    { name = "Lucky Umoru Zakari", email = "info@integra8ai.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "langgraph>=0.0.15",
    "pydantic>=2.0.0",
    "requests>=2.25.1",
    "python-dotenv>=0.19.0",
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langserve>=0.0.40",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src.agent"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
